Many changes whilst I had no internet, including:
Start unit testing find.c.
Use flock to stop simultaneous runs of burp_ca, hopefully preventing CA
corruption.
Unit tests for src/client/find.c, enabling the removal of some of the
time consuming end-to-end tests.
Don't use absolute path to 'sort' in 'make test'.
Fix bug where fs_name_max could get set to an incorrect value on the
server side, causing scandir to fail at the end of a backup.
Use /tmp/burp_ca.lock as the lockfile for burp_ca.
Fix protocol 1 phase 2 bug where librsync was never used.
Fix for a growing list of defunct child processes on busy servers, where
overlapping sigchlds would get conflated into one, and waitpid would only
get run once.
Protocol 2 phase 2 resume is working (but not for Windows in my CI for some
reason).
Make protocol 1 and 2 share resume code.
Unit tests for rabin/rconf and rabin/win.
Fix dpth_protocol1_set_from_string().
Set the server address in linux client burp.conf to "::1" if IPv6 is
supported, and "127.0.0.1" otherwise.
Sparse index merge testing.
Endian fixes.
Attempt to solve iobuf short read problem I saw once.
Fix bug where endfile marker was not always getting written in
phase2 protocol2 manifest.
Merge dindexes at the end of the backup.

+++ src/asfd.h
@@ -60,7 +60,7 @@ struct asfd
 	float ratelimit;
 	time_t rlstart;
 	int rlsleeptime;
-	unsigned long long rlbytes;
+	uint64_t rlbytes;
 
 	struct iobuf *rbuf;
